Always pass a String as the request body.
Before these changes a StringIO object _could_ be passed
into a resource but webmachine-ruby (AFAIK) *never* gives
you a IO object but always passes a String.

The result is that you expect a StringIO object when writing tests
but you expect a normal String when writing application code.

This pull request solves all that and ensures a String is always
given. FWIW you _can_ work around this by passing body: as a argument
to get/post/put/delete.